+++ src/service/mod.rs
@@ -40,6 +40,7 @@ pub mod signed_mail;
 pub mod spellcheck;
 pub mod tasks_api;
 pub mod update_check;
+pub mod update_download;
 pub mod vtimezone;
 
 pub use cache::CacheService;